requirements.txt
langchain
langchain_community  # Add if missing
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
openai
